The following android application is a revolutionary marketplace
The following android application is a AUEB project that delivers the marketplace service in a very innovative and revolutionary manner.
In order to differentiate ourselves from the common market apps, with the help of Machine learning in order to prevent unfitting ads to be uploaded and to help the category sorting and also the use of Google maps for enhanced information.